// ui-next/src/types/FormFieldTypes.ts
export enum UiIntegrationsFieldType {
  LLM_PROVIDER = "llmProvider",
  MODEL = "model",
  PROMPT_NAME = "promptName",
  TEXT = "text",
  VECTOR_DB = "vectorDB",
  NAMESPACE = "namespace",
  QUERY = "query",
  EMBEDDING_MODEL_PROVIDER = "embeddingModelProvider",
  EMBEDDING_MODEL = "embeddingModel",
  URL = "url",
  MEDIA_TYPE = "mediaType",
  DOC_ID = "docId",
  TEMPERATURE = "temperature",
  TOP_P = "topP",
  STOP_WORDS = "stopWords",
  MAX_TOKENS = "maxTokens",
  INDEX = "index",
  EMBEDDINGS = "embeddings",
  CHUNK_SIZE = "chunkSize",
  CHUNK_OVERLAP = "chunkOverlap",
  ID = "id",
  INSTRUCTIONS = "instructions",
  MESSAGES = "messages",
  MAX_RESULTS = "maxResults",
  JSON_OUTPUT = "jsonOutput",
  DIMENSIONS = "dimensions",
}